JJ-82 admin

+++ src/inners/use_cases/managements/AdminManagement.ts
@@ -218,16 +218,43 @@ export default class AdminManagement {
     )
   }
 
-  deleteOneById = async (id: string): Promise<Result<Admin | null>> => {
-    let deletedAdmin: Admin
+  deleteOneById = async (id: string, method: string): Promise<Result<Admin | null>> => {
     try {
-      const args: RepositoryArgument = new RepositoryArgument(
-        { id },
-        undefined,
-        undefined,
-        undefined
+      let deletedAdmin: Admin
+
+      const args: RepositoryArgument = new RepositoryArgument({ id })
+
+      if (method === 'hard') {
+        deletedAdmin = await this.adminRepository.deleteOne(args)
+      } else if (method === 'soft') {
+        const admin: Admin = await this.adminRepository.readOne(args)
+
+        const newAdmin: Admin = {
+          ...admin,
+          deletedAt: new Date()
+        }
+
+        const softDeleteArgs: RepositoryArgument = new RepositoryArgument(
+          { id },
+          undefined,
+          undefined,
+          newAdmin
+        )
+
+        deletedAdmin = await this.adminRepository.patchOne(softDeleteArgs)
+      } else {
+        return new Result<null>(
+          400,
+          'Invalid deletion method. Use "hard" or "soft".',
+          null
+        )
+      }
+
+      return new Result<Admin>(
+        200,
+        'Admin delete one by id succeed.',
+        deletedAdmin
       )
-      deletedAdmin = await this.adminRepository.deleteOne(args)
     } catch (error) {
       return new Result<null>(
         404,
@@ -235,10 +262,5 @@ export default class AdminManagement {
         null
       )
     }
-    return new Result<Admin>(
-      200,
-      'Admin delete one by id succeed.',
-      deletedAdmin
-    )
   }
 }
